requirements.txt
Django>=3.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.1.0
djangorestframework
psycopg2-binary
django-environ==0.8.1
flake8==4.0.1  
flake8-isort==4.1.1  
coverage==6.3.2  
black==22.3.0  
pylint-django==2.5.3  
pre-commit==2.18.1  
django-rest-passwordreset